enhance: [Ver.0.50] SingleUnitParameters を継承したクラスに変化があっても保存済みのデータに反映されるようにしたい #61
